+++ packages/app_usage/android/src/main/kotlin/dk/cachet/app_usage/Stats.java
@@ -1,4 +1,5 @@
-package cachet.plugins.appusage;
+package dk.cachet.app_usage;
+
 import android.app.usage.UsageStats;
 import android.app.usage.UsageStatsManager;
 import android.content.Context;