+++ packages/app_usage/example/test/widget_test.dart
@@ -13,14 +13,15 @@ import 'package:app_usage_example/main.dart';
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(new MyApp());
+    await tester.pumpWidget(MyApp());
 
     // Verify that platform version is retrieved.
     expect(
-        find.byWidgetPredicate(
-          (Widget widget) =>
-              widget is Text && widget.data.startsWith('Running on:'),
-        ),
-        findsOneWidget);
+      find.byWidgetPredicate(
+        (Widget widget) => widget is Text &&
+                           widget.data.startsWith('Running on:'),
+      ),
+      findsOneWidget,
+    );
   });
 }