+++ packages/app_usage/CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 1.0.0
+* Updated to the new Android API
+* Changed Android source code to Kotlin rather than Java
 
-* TODO: Describe initial release.
+## 0.9.0
+* Update example app to support Android X
+* Updated the README file to emphasize the need to use min SDK version 21 for Android
+
+## 0.0.1
+* Initial release
+